cluster_topic_modeling_emotions/requirements.txt

transformers
datasets
numpy
tqdm
scikit-learn
torch
matplotlib
seaborn
umap-learn
hdbscan
sentence-transformers
bertopic
plotly==5.13.1
openai
ollama
google-generativeai
